Restructure tests and move searcher class
- Move searcher into search namespace
- So we don't have to search for search
