+++ test/integration/elasticsearch_advanced_search_test.rb
@@ -1,7 +1,7 @@
 require "integration_test_helper"
 require "app"
 
-class LegacySearchAdvancedSearchTest < IntegrationTest
+class ElasticsearchAdvancedSearchTest < IntegrationTest
   def setup
     @index_name = "mainstream_test"
