+++ test/unit/elasticsearch_index_test.rb
