+++ test/unit/index_group_test.rb
