+++ test/unit/indexer/bulk_loader_test.rb
