+++ test/unit/indexer/index_queue_test.rb
