+++ test/unit/indexer/workers/amend_worker_test.rb
