+++ test/unit/indexer/workers/bulk_index_worker_test.rb
