+++ test/unit/indexer/workers/delete_worker_test.rb
