+++ test/unit/legacy_client/client_test.rb
