+++ test/unit/legacy_search/advanced_search_query_builder_test.rb
