+++ test/unit/legacy_search/elasticsearch_index_advanced_search_test.rb
