+++ test/unit/search/base_registry_test.rb
