+++ test/unit/search/best_bets_checker_test.rb
