+++ test/unit/search/elasticsearch_escaping_test.rb
