+++ test/unit/search/facet_example_fetcher_test.rb
