+++ test/unit/search/facet_option_test.rb
