+++ test/unit/search/matcher_set_test.rb
