+++ test/unit/search/presenters/entity_expander_test.rb
