+++ test/unit/search/presenters/highlighted_description_test.rb
