+++ test/unit/search/presenters/highlighted_title_test.rb
