+++ test/unit/search/presenters/result_presenter_test.rb
