+++ test/unit/search/query_components/core_query_test.rb
