+++ test/unit/search/query_components/facets_test.rb
