+++ test/unit/search/query_components/filter_test.rb
