+++ test/unit/search/query_components/highlight_test.rb
