+++ test/unit/search/query_components/popularity_test.rb
