+++ test/unit/search/query_components/text_query_test.rb
