+++ test/unit/search/result_set_test.rb
