+++ test/unit/search/search_builder_test.rb
