+++ test/unit/search/specialist_sector_registry_test.rb
