+++ test/unit/search/spell_check_fetcher_test.rb
