+++ test/unit/search/spell_check_presenter_test.rb
