+++ test/unit/search/suggestion_blacklist_test.rb
