+++ test/unit/search/timed_cache_test.rb
